format code

+++ src/makers/trident/TridentUnwindooor.sol
@@ -8,15 +8,12 @@ import "interfaces/Auth.sol";
 /// @notice Contract for breaking down LP positions
 /// @dev Built in mind for Trident LP positions (stable and constant product)
 contract TridentUnwindooor is Auth {
-  bytes4 private constant TRANSFER_SELECTOR = bytes4(keccak256(bytes('transfer(address,uint256)')));
+  bytes4 private constant TRANSFER_SELECTOR = bytes4(keccak256(bytes("transfer(address,uint256)")));
 
   error SlippageProtection();
   error TransferFailed();
 
-  constructor(
-    address _owner,
-    address user
-  ) Auth(_owner, user) {}
+  constructor(address _owner, address user) Auth(_owner, user) {}
 
   // Unwind pairs into a single token
   /// @dev burns pairs and unwinds position into one of the two tokens
@@ -62,13 +59,18 @@ contract TridentUnwindooor is Auth {
       IPool pair = IPool(pairs[i]);
       _safeTransfer(address(pair), address(pair), amounts[i]);
       bytes memory burnData = abi.encode(address(this), true);
-      
+
       IPool.TokenAmount[] memory withdrawnAmounts = pair.burn(burnData);
-      if (withdrawnAmounts[0].amount < minimumOut0[i] || withdrawnAmounts[1].amount < minimumOut1[i]) revert SlippageProtection();
+      if (withdrawnAmounts[0].amount < minimumOut0[i] || withdrawnAmounts[1].amount < minimumOut1[i])
+        revert SlippageProtection();
     }
   }
 
-  function _safeTransfer(address token, address to, uint value) internal {
+  function _safeTransfer(
+    address token,
+    address to,
+    uint256 value
+  ) internal {
     (bool success, bytes memory data) = token.call(abi.encodeWithSelector(TRANSFER_SELECTOR, to, value));
     if (!success || (data.length != 0 && !abi.decode(data, (bool)))) revert TransferFailed();
   }
